+++ src/trackedit/CMakeLists.txt
@@ -33,6 +33,8 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/internal/trackedituiactions.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/trackeditactionscontroller.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/trackeditactionscontroller.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/changedetection.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/changedetection.h
 
     ${CMAKE_CURRENT_LIST_DIR}/internal/au3/au3trackeditproject.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/au3/au3trackeditproject.h
