+++ src/trackedit/internal/changedetection.h
