+++ src/trackedit/tests/CMakeLists.txt
@@ -11,7 +11,7 @@ set(MODULE_TEST_SRC
     ${CMAKE_CURRENT_LIST_DIR}/mocks/trackeditprojectmock.h
 
     ${CMAKE_CURRENT_LIST_DIR}/au3interaction_tests.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/au3changedetection_tests.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/changedetection_tests.cpp
     ${CMAKE_CURRENT_LIST_DIR}/au3trackeditclipboard_tests.cpp
     )
